+++ mediagrains/numpy/__init__.py
@@ -18,6 +18,7 @@ Library for handling mediagrains in numpy arrays
 """
 
 from .videograin import VIDEOGRAIN, VideoGrain
+from .audiograin import AUDIOGRAIN, AudioGrain
 from . import convert  # noqa: F401
 
-__all__ = ['VideoGrain', 'VIDEOGRAIN']
+__all__ = ['VideoGrain', 'VIDEOGRAIN', 'AudioGrain', 'AUDIOGRAIN']